// types/auth/sanction.ts
export enum SanctionType {
  SANCTION_TYPE_UNSPECIFIED = 0,
  SUSPENSION = 1,
  WARNING = 2,
  BANNED = 3,
  UNRECOGNIZED = -1,
}

/** Message */
export interface Sanction {
  id: string;
  userId: string;
  sanctionType: SanctionType;
  reason?: string | undefined;
  startDate?: string | undefined;
  endDate?: string | undefined;
  createdAt: string;
}

/**
 * ---------------------
 * CRUD Responses
 * ---------------------
 */
export interface SanctionResponse {
  success: boolean;
  message?: string | undefined;
}

export interface AllSanctionsWithMetaResponse {
  data: Sanction[];
  meta: PaginationMeta | undefined;
}

export interface SingleSanctionResponse {
  message?: string | undefined;
  sanction: Sanction | undefined;
}

export interface DeleteSanctionResponse {
  success: boolean;
  message: string;
}

/**
 * ---------------------
 * CRUD Requests
 * ---------------------
 */
export interface GetAllSanctionsRequest {
  basicQuery: BasicQuery | undefined;
}

export interface GetSanctionRequest {
  sanctionId: string;
}

export interface CreateSanctionRequest {
  userId: string;
  sanctionType: SanctionType;
  reason?: string | undefined;
  startDate?: string | undefined;
  endDate?: string | undefined;
}

export interface UpdateSanctionRequest {
  sanctionId: string;
  sanctionType?: SanctionType | undefined;
  reason?: string | undefined;
  startDate?: string | undefined;
  endDate?: string | undefined;
}

export interface DeleteSanctionRequest {
  sanctionId: string;
}

export const SANCTIONS_PACKAGE_NAME = "sanctions";

/**
 * ---------------------
 * CRUD Service
 * ---------------------
 */

export interface SanctionServiceClient {
  createSanction(request: CreateSanctionRequest): Observable<SanctionResponse>;

  getSanction(request: GetSanctionRequest): Observable<SingleSanctionResponse>;

  getAllSanctions(request: GetAllSanctionsRequest): Observable<AllSanctionsWithMetaResponse>;

  updateSanction(request: UpdateSanctionRequest): Observable<SanctionResponse>;

  deleteSanction(request: DeleteSanctionRequest): Observable<DeleteSanctionResponse>;
}